repaired englishchords + relocated find_key
